promote nmagent retry to package, include final retry error, and adapt existing cni code
include last error cause when cooldown function errors (discuss) for compatability
use durations instead of millis
correct usage of number retries vs number of runs
create helper function for retrying in transparent vlan mode
create wrapper method that makes all passed in errors temporary (retriable) errors
