+++ retry/retry_example_test.go
@@ -1,4 +1,4 @@
-package internal
+package retry
 
 import (
 	"fmt"
